feat(observable):  Add support for importing local ojs/omd files
