+++ packages/observable-md/index.html
@@ -23,6 +23,18 @@
       .text(`\
 # Five-Minute Introduction
 
+\`\`\`
+import { mol } from "./samples/index.ojs";
+
+mol * 2;
+\`\`\`
+---
+\`\`\`
+import { nmol } from "./samples/index.omd";
+
+nmol * 2;
+\`\`\`
+
 \`\`\`ecl
 r := RECORD
   STRING20 Subject;
