+++ packages/observable-md/samples/index.ojs
@@ -1,0 +1,6 @@
+md`# This is OJS!`
+
+mol = mo + 2;
+
+import { mo } from "./lib.ojs"
+
