+++ packages/observable-md/samples/index.omd
@@ -1,0 +1,5 @@
+# This is OMD!
+
+```
+nmol = 40 + 4;
+```
