+++ packages/observable-md/samples/lib.ojs
@@ -1,0 +1,3 @@
+md `# lib.ojs`;
+
+mo = 38 + 40;
