+++ packages/observable-md/src/index.ts
@@ -1,7 +1,10 @@
 export * from "./__package__";
 
+export * from "./parsers";
 export * from "./omdRuntime";
 export * from "./ojsRuntime";
+export * from "./ojsModule";
+export * from "./ojsVariable";
 export * from "./domRender";
 export * from "./observable";
 export * from "./util";
