+++ tests/test-util/src/url.spec.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import { join } from "@hpcc-js/util";
 
 describe("url", function () {
-    it("join", function () {
+    it.only("join", function () {
         expect(join("aaa", "bbb", "ccc")).to.equal("aaa/bbb/ccc");
         expect(join()).to.equal("");
         expect(join("aaa")).to.equal("aaa");
@@ -10,5 +10,18 @@ describe("url", function () {
         expect(join("/aaa", "bbb", "ccc")).to.equal("/aaa/bbb/ccc");
         expect(join("/aaa/", "/bbb/", "/ccc/")).to.equal("/aaa/bbb/ccc");
         expect(join("aaa/", "/bbb/", "/ccc/")).to.equal("aaa/bbb/ccc");
+
+        expect(join("aaa", "../bbb/", "./ccc/")).to.equal("bbb/ccc");
+        expect(join("/aaa", "../bbb/", "./ccc/")).to.equal("/bbb/ccc");
+        expect(join("/aaa", "./bbb/", "./ccc/")).to.equal("/aaa/bbb/ccc");
+        expect(join("/aaa", "./bbb/", "../ccc/")).to.equal("/aaa/ccc");
+        expect(join("/aaa", "../bbb/", "../ccc/")).to.equal("/ccc");
+
+        expect(join("/aaa", "./bbb/", "../../ccc/")).to.equal("/ccc");
+
+        expect(join(".", "bbb", "ccc")).to.equal("bbb/ccc");
+        expect(join("/.", "bbb", "ccc")).to.equal("/bbb/ccc");
+        expect(join("./", "bbb", "ccc")).to.equal("bbb/ccc");
+        expect(join("/./", "bbb", "ccc")).to.equal("/bbb/ccc");
     });
 });
